ForeignKeyTypeChecker fail

+++ db/migrate/20110804175740_create_socializer_activity_objects.rb
@@ -3,7 +3,7 @@
 class CreateSocializerActivityObjects < ActiveRecord::Migration[7.1]
   def change
     create_table :socializer_activity_objects do |t|
-      t.integer  :activitable_id,             null: false
+      t.bigint   :activitable_id,             null: false
       t.string   :activitable_type,           null: false
       t.integer  :like_count,                 default: 0
       t.integer  :unread_notifications_count, default: 0
