+++ db/migrate/20110804184415_create_socializer_notes.rb
@@ -3,8 +3,8 @@
 class CreateSocializerNotes < ActiveRecord::Migration[7.1]
   def change
     create_table :socializer_notes do |t|
-      t.integer  :author_id, null: false
-      t.text     :content, null: false
+      t.bigint :author_id, null: false
+      t.text :content, null: false
 
       t.timestamps
     end
