+++ db/migrate/20110804192211_create_socializer_authentications.rb
@@ -3,10 +3,10 @@
 class CreateSocializerAuthentications < ActiveRecord::Migration[7.1]
   def change
     create_table :socializer_authentications do |t|
-      t.integer  :person_id, null: false
-      t.string   :provider,  null: false
-      t.string   :uid,       null: false
-      t.string   :image_url
+      t.bigint :person_id, null: false
+      t.string :provider, null: false
+      t.string :uid, null: false
+      t.string :image_url
 
       t.timestamps
     end
