+++ db/migrate/20110805085417_create_socializer_circles.rb
@@ -3,7 +3,7 @@
 class CreateSocializerCircles < ActiveRecord::Migration[7.1]
   def change
     create_table :socializer_circles do |t|
-      t.integer :author_id, index: true, null: false
+      t.bigint :author_id, index: true, null: false
       t.string :display_name, null: false
       t.text :content
