+++ db/migrate/20110805094846_create_socializer_comments.rb
@@ -3,8 +3,8 @@
 class CreateSocializerComments < ActiveRecord::Migration[7.1]
   def change
     create_table :socializer_comments do |t|
-      t.integer  :author_id, null: false
-      t.text     :content,   null: false
+      t.bigint :author_id, null: false
+      t.text :content, null: false
 
       t.timestamps
     end
